Fix N+1 query performance in inventory counting endpoints
Add bulk query methods (BulkItemCheckAsync, BulkBinCheckAsync,
BulkGetBinCodesAsync) that fetch all data upfront with batched IN
clauses (max 500 params per batch) instead of querying per item/bin
in a loop. Applied to GetCountingSummaryReport, GetCountingContent,
and PrepareCountingData — reducing ~3N-4N sequential SAP round-trips
to 3-4 total queries.

Also fix pre-existing decimal-to-double cast errors in
CountingCreation.cs and PickingUpdate.cs.

+++ Adapters.Windows/SBO/Helpers/CountingCreation.cs
@@ -88,7 +88,7 @@ public class CountingCreation(
                     line.WarehouseCode   = whsCode;
                     line.BinEntry        = countedBin.BinEntry;
                     line.Counted         = BoYesNoEnum.tYES;
-                    line.CountedQuantity = countedBin.CountedQuantity;
+                    line.CountedQuantity = (double)countedBin.CountedQuantity;
                     totalLines++;
                     
                     logger.LogDebug("Added counting line for item {ItemCode} in bin {BinEntry} with quantity {Quantity} (system: {SystemQuantity})", 
@@ -101,7 +101,7 @@ public class CountingCreation(
                 line.ItemCode        = value.Value.ItemCode;
                 line.WarehouseCode   = whsCode;
                 line.Counted         = BoYesNoEnum.tYES;
-                line.CountedQuantity = value.Value.CountedQuantity;
+                line.CountedQuantity = (double)value.Value.CountedQuantity;
                 totalLines++;
                 
                 logger.LogDebug("Added counting line for item {ItemCode} with quantity {Quantity} (system: {SystemQuantity}, variance: {Variance})",
